fixed broken files
Jekyll seems to get very confused about : in titles. It makes the posts always be the newest post and have no title. That behavior is very bad because those posts will get pinned.

This commit fixes the issue by replacing : wich &#58; which renders the same.

If you think this is not an issue: Just take a look at the current maschinendeck.org.